+++ changelog.md
@@ -19,6 +19,7 @@
     - Fix creating duplicate function stubs in generated code
     - Fix collection dialog editor not comparing arrays correctly
     - Fix issue when trying to read include files for peek/goto define
+    - Area Designer: Fix type in monster resistances code
 - **Changed:**
   - Update electron 31.1.0 to 32.1.0
   - Update better-sqlite3 11.2.1 to 11.3.0